# Make 32-bit watershed function