--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools >= 61.0", "wheel", "Cython"]
+requires = ["setuptools >= 61.0", "wheel", "numpy < 2", "Cython < 3"]
 build-backend = "setuptools.build_meta"
 
 [project]