Extracts the `watershed` function from scikit-image to make a 32-bit Cython version alongside 64-bit. This prevents an upscaling to 64-bit to be compatible with the compiled binary (which now supports both 32 & 64 bits).

Along the way … bundle in an `in_place` option that controls whether the code forces copies before making destructive changes. Since we don't need to hold on to the original inputs, we can save allocations by e.g. multiplying in-place rather than creating a new output array then reassigning to the input variable.

## Benchmarking

Use `memray` for memory benchmarking. https://github.com/bloomberg/memray

`tracemalloc` doesn't see the numpy native C allocs, which are, well, the whole point of this.

`memory_profiler` is great to see peak usage, however that is sensitive to when the garbage collector happens to run (which would happen more often for larger inputs). It's also sampled every 0.1s. The peak usage tells us how much was in memory at once, which is not necessarily how much *had* to be in memory at once (the GC might or might not have cleaned up unused memory yet).

### Input

9777 * 14239 pixels (human breast cancer from 10x genomics)
~139 M pixels

times 2 predictions/pixel
~278 M

times 32bits (4 bytes) per pixel
~1 GB

Or ~0.5GB per model prediction.

<img width="314" alt="Screenshot 2024-12-01 at 3 59 11 PM" src="https://github.com/user-attachments/assets/6fe75289-18ff-4006-a74d-85092c1a0120">

### Status quo

<img width="313" alt="Screenshot 2024-12-01 at 4 04 32 PM" src="https://github.com/user-attachments/assets/cb449ad3-1c49-4556-bf0c-03c1a8ce8fa6">

### 32-bit, still copying

We drop by half of the input, because we're no longer doubling one of the input predictions into 64 bits.

<img width="313" alt="Screenshot 2024-12-01 at 4 05 07 PM" src="https://github.com/user-attachments/assets/f43d1004-c778-4053-82b4-1a5e60938c73">

### 32-bit, in-place

We drop by a whole input this time, although I think it's actually one of the predictions that gets upsampled to 64-bit by h_maxima (a separate issue).

<img width="313" alt="Screenshot 2024-12-01 at 4 05 23 PM" src="https://github.com/user-attachments/assets/fa14f2aa-31a9-4ea3-b950-03db297bf562">

### Benchmark summary

This is total memory for `main` (including the input).

| Input \ Memory usage | Before | After | Change |
|--|---|---|--|
| ~139M pixels | 9.1 GB | 7.5 GB | -1.6 GB (18%) |

I believe this effectively means saving `4*pixels` bytes of memory.

In principle, for a 3B pixel image, we're saving ~12 GB of memory.

Benchmark files:

[postprocess-memory-benchmarks.zip](https://github.com/user-attachments/files/17970872/postprocess-memory-benchmarks.zip)

----

Solves #377